Add files via upload
This ML model predict duplicate question. If there are multiple numbers of questions of similar type then it recognizes. So it becomes useful for providing answer of all the similar kind of question by same answer.
Example we can say for Quora. On Quora many people ask questions if model find there similar kind of question it will be easy to answer.